Updated UserBusinessImpl to use JPA repo instead of Hibernate DAOs

+++ src/main/java/org/randomcoder/bo/UserBusiness.java
@@ -122,5 +122,5 @@ public interface UserBusiness
 	 * 
 	 * @return count of user objects
 	 */
-	public int countUsers();
+	public long countUsers();
 }
